requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
gitdb==4.0.12
GitPython==3.1.46
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.14
langchain-core==1.2.24
langchain-groq==1.1.2
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.19.0
nest-asyncio==1.6.0
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
